requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.1
